project/requirements.txt
Flask
Flask-PyMongo
Flask_restful
requests 
pytest
flaskr
